+++ ruoyi-admin/src/main/java/org/dromara/system/domain/bo/WaterMonitoringDataBo.java
@@ -36,17 +36,17 @@ public class WaterMonitoringDataBo extends BaseEntity {
     /**
      * 水流量（L/min）
      */
-    private Long flowRate;
+    private Double flowRate;
 
     /**
      * 水压（MPa）
      */
-    private Long pressure;
+    private Double pressure;
 
     /**
      * 水质指数（0-100）
      */
-    private Long qualityIndex;
+    private Double qualityIndex;
 
     /**
      * 记录时间
